Move run button to TestLogs
